feat: user can delete  a board
