+++ client/store/user.js
@@ -7,6 +7,7 @@ import history from '../history'
 const GET_USER = 'GET_USER'
 const REMOVE_USER = 'REMOVE_USER'
 const SET_USER_BOARDS = 'SET_USER_BOARDS'
+const REMOVE_BOARD = 'REMOVE_BOARD'
 
 /**
  * INITIAL STATE
@@ -19,6 +20,11 @@ const defaultUser = {}
 const getUser = user => ({type: GET_USER, user})
 const removeUser = () => ({type: REMOVE_USER})
 const setUserBoards = boards => ({type: SET_USER_BOARDS, boards})
+
+export const removeBoard = boardId => ({
+  type: REMOVE_BOARD,
+  boardId
+})
 
 /**
  * THUNK CREATORS
@@ -67,6 +73,17 @@ export const logout = () => async dispatch => {
   }
 }
 
+export const deleteBoard = boardId => {
+  return async dispatch => {
+    try {
+      await axios.delete(`/api/projects/${boardId}`)
+      dispatch(removeBoard(boardId))
+    } catch (error) {
+      console.log('Error deleting board', error)
+    }
+  }
+}
+
 /**
  * REDUCER
  */
@@ -78,6 +95,10 @@ export default function(state = defaultUser, action) {
       return defaultUser
     case SET_USER_BOARDS:
       return {...state, savedBoards: action.boards}
+    case REMOVE_BOARD: {
+      let filt = state.savedBoards.filter(board => board._id !== action.boardId)
+      return {...state, savedBoards: filt}
+    }
     default:
       return state
   }
